feat: working but not what i want
