+++ src/components/Calendar/Heading.tsx
@@ -1,11 +1,7 @@
 import React, { FC } from 'react';
-import { HeadingProps } from './types';
-import clsx from 'clsx';
+import { CompoundComponentProps } from './types';
 
-export const Heading: FC<HeadingProps> = ({ children, className }) => {
-  return (
-    <div className={clsx('flex items-center justify-between p-4', className)}>
-      {children}
-    </div>
-  );
-};
+export const Heading: FC<CompoundComponentProps> = ({
+  className,
+  children,
+}) => <div className={className}>{children}</div>;
