+++ src/components/Calendar/index.ts
@@ -1,2 +1,9 @@
-export { default as Calendar } from './Calendar';
+export { Calendar } from './Calendar';
+export { Grid, GridHeader, GridBody } from './Grid';
+export { Navigation } from './NavigationButtons';
+export { MonthSelect, YearSelect } from './Selects';
+export { Legend, LegendItem } from './Legend';
+export { Messaging } from './Messaging';
+export { Heading } from './Heading';
+export { Caption } from './Caption';
 export type { CalendarProps } from './types';
